+++ src/main/java/org/http/Releaseable.java
@@ -1,0 +1,15 @@
+package org.http;
+
+/**
+ * 可释放的声明
+ * @author zouziwen
+ *
+ */
+public interface Releaseable {
+	
+	/**
+	 * 释放资源
+	 */
+	void release(); 
+	
+}